refactoring more
