+++ src/codefresh/runtime-type.ts
@@ -1,5 +1,5 @@
 export enum RuntimeType {
   pipelines = 'Pipelines Runtime',
   gitops = 'GitOps Runtime',
-  onprem = 'On-Prem'
+  onprem = 'On-Prem',
 }
